pyproject.toml
[tool.poetry]
name = "ridi-to-read"
version = "0.1.0"
description = ""
authors = ["Yoon Oh <yoonoh930@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
